update as per requirements

+++ database/migrations/2024_10_17_085835_create_t_suppliers_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('t_suppliers', function (Blueprint $table) {
             $table->id();
-            $table->integer('supplier_id');
+            $table->unsignedBigInteger('supplier_id');
             $table->string('name')->unique();
             $table->string('address_line_1');
             $table->string('address_line_2');
